Etapa02

+++ Servidor/src/main/java/com/projetosd/grpc/resources/Input.java
@@ -1,12 +1,14 @@
 package com.projetosd.grpc.resources;
 
+import java.math.BigInteger;
+
 public class Input {
-	private long id;
+	private BigInteger id;
     private String content;
     private int operation; // 0 - Insert, 1 - Select, 2 - Update, 3 - Delete;
     private EventSource eventSource;
 
-    public Input(long _id, String _content, int _operation, EventSource _eventSource) {
+    public Input(BigInteger _id, String _content, int _operation, EventSource _eventSource) {
     	this.id = _id;
         this.content = _content;
         this.operation = _operation;
@@ -14,17 +16,17 @@ public class Input {
     }
     
     public Input() {
-    	this.id = -1;
+    	this.id = BigInteger.valueOf(-1);
     	this.content = "";
     	this.operation = -1;
     	this.eventSource = null;
     }
     
-    public void setId(long _id) {
+    public void setId(BigInteger _id) {
     	this.id = _id;
     }
     
-    public long getId() {
+    public BigInteger getId() {
     	return this.id;
     }
